move ui related to journal to own files
